Finished Independent Project
Enhancements: Made the gun rotate and made it shoot bullets at the direction it is facing.
Made the enemies move toward the player.
Fixed some collision bugs with colliding with objects more than one time.
The enemies and obstacles can now damage the player if they collide with it.
Also, added a boss, a victory method, game over method, and a lot of UI elements. There is also 2 requirements to win.  (No Citations Needed)
Have Fun!